+++ examples/ContainerConfig.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Zen\Examples;
 
-use Interop\Container\ContainerInterface;
+use Psr\Container\ContainerInterface;
 use WoohooLabs\Zen\Config\AbstractContainerConfig;
 use WoohooLabs\Zen\Config\EntryPoint\WildcardEntryPoint;
 use WoohooLabs\Zen\Config\Hint\DefinitionHint;