+++ src/Config/ContainerConfigInterface.php
@@ -3,8 +3,8 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Zen\Config;
 
-use WoohooLabs\Zen\Config\Hint\DefinitionHintInterface;
 use WoohooLabs\Zen\Config\EntryPoint\EntryPointInterface;
+use WoohooLabs\Zen\Config\Hint\DefinitionHintInterface;
 
 interface ContainerConfigInterface
 {