+++ src/Exception/ContainerException.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 namespace WoohooLabs\Zen\Exception;
 
 use Exception;
-use Interop\Container\Exception\ContainerException as InteropContainerException;
+use Psr\Container\ContainerExceptionInterface as PsrContainerExceptionInterface;
 
-class ContainerException extends Exception implements InteropContainerException
+class ContainerException extends Exception implements PsrContainerExceptionInterface
 {
 }