+++ tests/unit/Fixture/Container/ContainerWithEntryPoint.php
@@ -10,7 +10,7 @@ class ContainerWithEntryPoint extends AbstractContainer
      */
     protected $entryPoints = [
         \WoohooLabs\Zen\Tests\Unit\Fixture\Container\ContainerWithEntryPoint::class => 'WoohooLabs__Zen__Tests__Unit__Fixture__Container__ContainerWithEntryPoint',
-        \Interop\Container\ContainerInterface::class => 'Interop__Container__ContainerInterface',
+        \Psr\Container\ContainerInterface::class => 'Psr__Container__ContainerInterface',
         \WoohooLabs\Zen\Tests\Unit\Double\StubDefinition::class => 'WoohooLabs__Zen__Tests__Unit__Double__StubDefinition',
     ];