+++ tests/unit/Utils/FileSystemUtilTest.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 namespace WoohooLabs\Zen\Tests\Unit\Utils;
 
 use PHPUnit\Framework\TestCase;
-use WoohooLabs\Zen\Tests\Unit\Fixture\DependencyGraph\EntryPoint\EntryPointGAbstract;
 use WoohooLabs\Zen\Tests\Unit\Fixture\DependencyGraph\EntryPoint\EntryPointA;
 use WoohooLabs\Zen\Tests\Unit\Fixture\DependencyGraph\EntryPoint\EntryPointC1;
 use WoohooLabs\Zen\Tests\Unit\Fixture\DependencyGraph\EntryPoint\EntryPointC2;
 use WoohooLabs\Zen\Tests\Unit\Fixture\DependencyGraph\EntryPoint\EntryPointEInterface;
+use WoohooLabs\Zen\Tests\Unit\Fixture\DependencyGraph\EntryPoint\EntryPointGAbstract;
 use WoohooLabs\Zen\Tests\Unit\Fixture\DependencyGraph\EntryPointD1\EntryPointD1;
 use WoohooLabs\Zen\Tests\Unit\Fixture\DependencyGraph\EntryPointD2\EntryPointD2;
 use WoohooLabs\Zen\Utils\FileSystemUtil;